rag/multimodal-rag/requirements.txt
transformers>=4.30.0
torch>=2.0.0
pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
matplotlib>=3.5.0

